+++ src/assets/icons/Icons.jsx
@@ -8,7 +8,6 @@ import {
   AiOutlineEye,
   AiOutlineEyeInvisible,
 } from "react-icons/ai";
-import { AiOutlineDelete, AiOutlineUserDelete } from "react-icons/ai";
 
 export {
   BiSearch,
