+++ src/components/index.jsx
@@ -1,0 +1,2 @@
+import { SideNav } from "./SideNav";
+export { SideNav };
